requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.5
django-better-admin-arrayfield==1.4.2
django-celery-beat==2.7.0
django-ckeditor==6.7.2
django-environ==0.12.0
django-filter==24.3
django-image-uploader-widget==1.0.1
django-js-asset==3.0.1
django-modeltranslation==0.19.12
django-mptt==0.16.0
django-phonenumber-field==8.0.0
django-timezone-field==7.1
djangorestframework==3.15.2
drf-yasg==1.21.8
h11==0.14.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
outcome==1.3.0.post0
packaging==24.2
path==17.1.0
phonenumbers==8.13.54
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sorl-thumbnail==12.11.0
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
